requirements.txt
app==0.0.1
argon2-cffi==18.3.0
asn1crypto==0.24.0
astroid==2.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
blessed==1.15.0
cached-property==1.5.1
cement==2.8.2
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
diff-match-patch==20121119
dj-database-url==0.5.0
Django==1.11.17
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-csv-export-view==0.5.0
django-debug-toolbar==1.10.1
django-excel==0.0.4
django-excel-response==2.0.3
django-extensions==2.1.5
django-formtools==2.1
django-heroku==0.3.1
django-import-export==1.0.1
django-multiforloop==0.2.1
django-qr-code==1.0.0
django-screenshots==0.1.7
django-widget-tweaks==1.4.2
djangorestframework==3.9.1
dockerpty==0.4.1
docopt==0.6.2
drfdocs==0.0.11
et-xmlfile==1.0.1
Faker==1.0.2
Flask==0.10.1
Flask-MySQL==1.4.0
future==0.16.0
# get==2018.11.19
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
lml==0.0.4
lxml==4.2.5
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.15.2
odfpy==1.3.6
openpyxl==2.5.9
pandas==0.23.4
passlib==1.7.1
Pillow==5.2.0
# post==2018.11.20
psycopg2==2.7.5
public==2018.11.20
pycparser==2.19
pyexcel==0.5.9.1
pyexcel-io==0.5.9.1
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
PyJWT==1.7.1
pylint==2.2.2
PyMySQL==0.9.2
PySocks==1.6.8
python-dateutil==2.7.3
python-dotenv==0.10.1
pytz==2018.5
PyYAML==3.13
qrcode==6.0
query-string==2018.11.20
# request==2018.11.20
requests==2.19.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
termcolor==1.1.0
text-unidecode==1.2
texttable==0.9.1
twilio==6.10.1
typed-ast==1.3.1
unicodecsv==0.14.1
urllib3==1.23
virtualenv==16.0.0
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.14.1
whitenoise==3.3.1
wrapt==1.11.1
xadmin==0.6.1
xlrd==1.1.0
XlsxWriter==1.1.1
xlwt==1.3.0
